requirements.txt
Django==5.0.7
django-jazzmin==3.0.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1